+++ Pandora/Paradigm/Primary/Functor/Edges.hs
@@ -2,9 +2,9 @@ module Pandora.Paradigm.Primary.Functor.Edges where
 
 import Pandora.Pattern.Category (($))
 import Pandora.Pattern.Functor.Covariant (Covariant ((-<$>-)))
-import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Traversable (Traversable ((<<-)))
 import Pandora.Paradigm.Primary.Algebraic.Exponential ()
+import Pandora.Paradigm.Primary.Algebraic (point_)
 
 data Edges a = Empty | Leap a | Connect a | Overlay a
 
@@ -15,7 +15,7 @@ instance Covariant Edges (->) (->) where
 	f -<$>- Leap x = Leap $ f x
 
 instance Traversable Edges (->) (->) where
-	_ <<- Empty = point Empty
+	_ <<- Empty = point_ Empty
 	f <<- Connect x = Connect -<$>- f x
 	f <<- Overlay x = Overlay -<$>- f x
 	f <<- Leap x = Leap -<$>- f x