+++ CHANGELOG.md
@@ -561,3 +561,4 @@
 * Define `sum` method
 * Add `Monoidal` constraint to `Applicative` and `Alternative` constraints
 * Define `point_` method
+* Use `Monoidal` superclass in `Traversable` instead of `Pointable`